Fixing prettier/lint problems

+++ DenialofService.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.3;
 
 contract KingOfEther {
     address public king;
-    uint public balance;
+    uint256 public balance;
 
     function claimThrone() external payable {
         require(msg.value > balance, "Need to pay more to become the king");
@@ -22,7 +22,6 @@ contract Attack {
     constructor(KingOfEther _kingOfEther) {
         kingOfEther = KingOfEther(_kingOfEther);
     }
-
 
     function attack() public payable {
         kingOfEther.claimThrone{value: msg.value}();
