+++ Reentrancy.sol
@@ -1,14 +1,15 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.3;
+
 contract EtherStore {
-    mapping(address => uint) public balances;
+    mapping(address => uint256) public balances;
 
     function deposit() public payable {
         balances[msg.sender] += msg.value;
     }
 
     function withdraw() public {
-        uint bal = balances[msg.sender];
+        uint256 bal = balances[msg.sender];
         require(bal > 0);
 
         (bool sent, ) = msg.sender.call{value: bal}("");
@@ -18,7 +19,7 @@ contract EtherStore {
     }
 
     // Helper function to check the balance of this contract
-    function getBalance() public view returns (uint) {
+    function getBalance() public view returns (uint256) {
         return address(this).balance;
     }
 }
@@ -44,7 +45,7 @@ contract Attack {
     }
 
     // Helper function to check the balance of this contract
-    function getBalance() public view returns (uint) {
+    function getBalance() public view returns (uint256) {
         return address(this).balance;
     }
 }
