+++ Vault.sol
@@ -2,26 +2,26 @@
 pragma solidity ^0.8.3;
 
 contract Vault {
-    uint public count = 123;
+    uint256 public count = 123;
     address public owner = msg.sender;
     bool public isTrue = true;
     uint16 public u16 = 31;
     bytes32 private password;
 
     // constants do not use storage
-    uint public constant someConst = 123;
+    uint256 public constant someConst = 123;
 
     // slot 3, 4, 5 (one for each array element)
     bytes32[3] public data;
 
     struct User {
-        uint id;
+        uint256 id;
         bytes32 password;
     }
 
     User[] private users;
 
-    mapping(uint => User) private idToUser;
+    mapping(uint256 => User) private idToUser;
 
     constructor(bytes32 _password) {
         password = _password;
@@ -35,14 +35,19 @@ contract Vault {
     }
 
     function getArrayLocation(
-        uint slot,
-        uint index,
-        uint elementSize
-    ) public pure returns (uint) {
-        return uint(keccak256(abi.encodePacked(slot))) + (index * elementSize);
+        uint256 slot,
+        uint256 index,
+        uint256 elementSize
+    ) public pure returns (uint256) {
+        return
+            uint256(keccak256(abi.encodePacked(slot))) + (index * elementSize);
     }
 
-    function getMapLocation(uint slot, uint key) public pure returns (uint) {
-        return uint(keccak256(abi.encodePacked(key, slot)));
+    function getMapLocation(uint256 slot, uint256 key)
+        public
+        pure
+        returns (uint256)
+    {
+        return uint256(keccak256(abi.encodePacked(key, slot)));
     }
 }
